Add basic support for Flask-specific import of extensions
It works only for extensions that switched to `flask_foo` schema, it
does not work for extensions like `flaskfoo`. Moreover completion works
only "inside" extensions, it does not let you complete extension itself.

So the in following statement `foo` will not be suggested:

    from flask.ext.fo

However, completion should work in other cases, such as:

    from flask.ext.foo import
    from flask.ext.foo.bar import
    import flask.ext.foo.